Temporary fix calculator(extension use for CORS)
