+++ .gitignore
@@ -25,4 +25,3 @@ dist-ssr
 *.sln
 *.sw?
 node_modules/
-/package-lock.json
