+++ React-Calculator/src/main.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import Calculator from './components/Calculator.jsx';
 import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <App />
+    <Calculator />
+    
   </React.StrictMode>,
 );
